\name{BayesEM-package}
\alias{BayesEM-package}
\alias{BayesEM}
\docType{package}
\title{
\packageTitle{BayesEM}
}
\description{
\packageDescription{BayesEM}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{BayesEM}
\packageIndices{BayesEM}

}
\author{
\packageAuthor{BayesEM}

Maintainer: \packageMaintainer{BayesEM}
}
\references{
	Submitted Manuscript 
}

\keyword{ package }

