\name{BayesEM-package}
\alias{BayesEM-package}
\alias{BayesEM}
\docType{package}
\title{
\packageTitle{BayesEM}
}
\description{
\packageDescription{BayesEM}
}
\details{

}
\author{
\packageAuthor{BayesEM}

Maintainer: \packageMaintainer{BayesEM}
}
\references{
~~ Literature or other references for background information ~~
}
